backend/requirements.txt
fastapi>=0.103.0
uvicorn>=0.23.0
-e ../../fastapi-payments
# -e ./fastapi-payments
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
faststream[memory]>=0.2.0
pydantic>=2.0.0
email-validator>=2.0.0
stripe>=6.0.0